disable eslint for packages/app/index.js
Since this file is autogenerated from expo it doesn't behave our eslint rules
to fix this, we were running a postinstall script which corrected the
formatting of the file. This works at the root level, but when running yarn install
at the package level, the file is still generated, but the root level postinstall
is never triggered. No point battling against this autogenerated file with lint rules