Refactored @withorbit/store into multiple packages to avoid issues with conflicting web / RN types
This commit breaks @withorbit/store into @withorbit/store-fs and @withorbit/store-web, with @withorbit/store-shared containing platform-independent code. The motivation here is in challenges which arise when code which depends on both DOM types and React Native types in the same module.

In particular, if you add `dom` to the list of included libraries in `tsconfig.json`, numerous types in that definition file will conflict with types declared in React Native's types file. Of course, `OrbitStoreWeb` needs the `dom` library definitions (for IDB types), and `OrbitStoreFS.native` neesd the RN library definitions.

`@withorbit/app` gets around this problem by adding `skipLibCheck: true` to its TS configuration, which prevents TS from type-checking the libraries it uses. I don't know how to avoid it in the case of the app, which tightly mingles web-specific and RN-specific code, but I don't want to propagate that kludge to other libraries, particularly low-level ones like these. If one of our dependencies changes its exported types in a way which causes errors, we want to know about it!

This change is mostly for the worse, I think. But it does make the structure of the individual modules somewhat more straightforward, and it's nice that `store-fs` and `store-web` have only a single default public export.